KAN-54 refactor: move store and API base logic from app to shared
