--- src/app/api/index.ts
@@ -1,2 +1,0 @@
-export { baseQuery } from './base-query'
-export { baseApi } from './base-api'
