--- src/app/model/index.ts
@@ -1,1 +1,0 @@
-export * from './store'
