+++ src/shared/store/base-api.ts
@@ -1,5 +1,5 @@
 import { createApi } from '@reduxjs/toolkit/query/react'
-import { baseQuery } from '.'
+import { baseQuery } from '@/shared/store/base-query'
 
 export const baseApi = createApi({
    reducerPath: 'baseApi',
