+++ src/shared/store/index.ts
@@ -1,0 +1,3 @@
+export { store, type RootState, type AppDispatch } from './store'
+export { baseApi } from './base-api'
+export { baseQuery } from './base-query'
