+++ src/shared/store/providers/StoreProvider.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import { Provider } from 'react-redux'
-import { store } from '@/app/model'
 import { ReactNode } from 'react'
+import { store } from '..'
 
 export const StoreProvider = ({ children }: { children: ReactNode }) => {
    return <Provider store={store}>{children}</Provider>
